+++ src/tko/game/game.py
@@ -78,7 +78,7 @@ class Game:
 
     def load_sources(self, sources: list[RepSource], language: str):
         for source in sources:
-            filename = source.target_path
+            filename = source.get_file_path()
             filters = source.filters
             content = ""
             if filename == "":